split 'match -E' into its own ematch() function
Not parsing an -E option makes it possible to deal with empty removal, which
helps eliminate the need to quote variables under 'use safe'. The text to match
can now be empty (but not the regex argument). That means one argument is assumed
to mean that the text argument is empty and removed, so one argument always
returns false instead of producing an error.