Complete integrating file tests into is() and can()
This commit completes the work started in commits 7e38e73, 9cc3664 and
9cc3664 on the new (incompatible) 0.1 branch to create a more grammatically
robust set of file type and permission tests that is completely tolerant of
empty removal (which is essential for using unquoted variables under 'use
safe').

Consider the old way: "isdir -L $var", what is $var is empty? This will then
revert to "isdir -L" which returns true if a directory named "-L" exists,
instead of false. The new way: "is -L dir $var" avoids this problem: the
empty removal is correctly derected and it always correctly returns false.

So, by letting the -L option precede a keyword instead of the filename,
there can now never be any confusion between the -L option and a file named
"-L".

An additional advantage is that integrating these into a just two functions,
is() and can(), reduces clutter in the main shell function namespace.

	modified:   bin/modernish
	modified:   install.sh
	modified:   libexec/modernish/safe.mm
	modified:   libexec/modernish/sys/baseutils.mm
	modified:   libexec/modernish/sys/dirutils.mm
	modified:   libexec/modernish/sys/textfiles.mm
	modified:   libexec/modernish/sys/trace.mm
	modified:   libexec/modernish/sys/userinfo.mm
	modified:   share/doc/modernish/WRITEUP.md
	modified:   share/doc/modernish/examples/copy-timestamps.sh
	modified:   share/doc/modernish/examples/get-firefox.sh
	modified:   share/doc/modernish/examples/resurrectjoe-1.0.3/resurrectjoe
	modified:   share/doc/modernish/examples/testshells.sh
	modified:   share/doc/modernish/testsuite/loopspeed/loopspeed-test
	modified:   share/doc/modernish/testsuite/traverse-test
	modified:   uninstall.sh