New which() option: -P<num>: Strip pathname elements from right. This is helpful for determining the install prefix for an installed package.
For instance, on my current system, 'which -a make' outputs:

	/Applications/Xcode.app/Contents/Developer/usr/bin/make
	/usr/bin/make

The modernish command 'which -aP2 make' strips the last two elements of each
path, yielding:

	/Applications/Xcode.app/Contents/Developer/usr
	/usr

which tells you where the associated files for each 'make' are,
for instance the man page in each prefix plus share/man/man1/make.1

This feature was written because uninstall.sh required it, so uninstall.sh
is updated to take advantage of this.

Update WRITEUP.md with this new feature.