uninstall.sh: don't try to remove the main destdir even if empty
For instance, if modernish is test-installed under /var/tmp and uninstall.sh
is run to remove it, and /var/tmp was empty before modernish was installed
in it, then the uninstaller would try to remove /var/tmp after removing
modernish. If not root then this would kill the uninstaller (due to 'harden
rmdir'), but if root this could have caused the loss of /var/tmp.