trap stack: no longer ignore signals trapped by 'pushtrap' and not 'trap'
Traps set with 'pushtrap' have to work a bit differently if they are not to
interfere with each other, as well as with the POSIX 'trap' command (which
was intercepted to play nice with the trap stack, emulating POSIX
behaviour).

Before this commit, using 'pushtrap' with a real signal (i.e. not EXIT)
would cause that signal to be handled and then ignored, as a builtin system
trap was set to handle it and that's what 'trap' normally does. Even using
the emulated POSIX command 'trap - SIGNAL' to deactivate that trap would
still not cause the signal to stop being ignored. This made the trap stack
incompatible with normal POSIX trap operation.

This commit should fix that. If there are only stack traps and no POSIX trap
for a particular signal, then the stack traps are executed, the signal is
un-ignored (i.e. the system trap is unset), and a 'kill' is sent to self
("$$") to kill the program using the expected signal. This ought to make the
emulated 'trap' command behave exactly as expected while still playing nice
with the trap stack.

  [Note: Traps set with 'pushtrap' should never exit the shell, because the
   whole point of them is that they shouldn't interfere with each other.
   Since modernish modules can use 'pushtrap' to clean up after themselves,
   you don't necessarily know what other traps are pushed.]