zsh: disable buggy test/[ builtin; make BUG_APPENDC test zsh-only
On older zsh, the test/[ builtin is too buggy to use. Disable it if found
buggy, so the external equivalent (which is mandated by POSIX) is used
instead.

Also, zsh-specific bug testing tweaks.

No longer test for BUG_APPENDC on non-zsh shells as the test is too
expensive and that bug is extremely unlikely ever to appear on other POSIXy
shells. (So far that's the first time I'm doing any shell-specific test...
and hopefully the last time... ugh)