trap stack: don't unignore singals on interactive shells
Unignoring signals (i.e. unsetting their POSIX trap) causes funny behaviour
on interactive shells. Some shells (e.g. dash) don't exit but immediately
return to the command prompt, orphaning the trap stack from its POSIX trap
containing the _Msh_doTraps command. So first check if shell is interactive.

Also, remove a shell bug workaround that turned out to be unnecessary. The
bug seemed to be that the resent signal did not have an effect on zsh and
older bash. In reality these shells merely ignore the resent signal for the
duration of the trap routine (in this case, _Msh_doTraps()) and act on it
at the end thereof. So if the 'kill' to resend the signal is the last
command in the function, all is good on every shell.