mktemp(): new option -C (push cleanup trap); whitespace validation
New option -C: Automated cleanup. Push a trap to remove the files on exit
using the modernish trap stack.

Validate whitespace: mandate the use of the -Q option (shell-quoted output)
if there is more than one template and at least one of them contains
whitespace, such as spaces or newlines. (Technically this should only have
to be a check on newlines as the newline is the only separator on
non-shellquoted output, but this seems safer...)

Remove extra trailing space from -Q output.