[Incompatible change] Unload arith extras into modules
The following arithmetic functions are nice syntactic sugar, but essentially
superfluous, because 'let', which is a builtin on many shells and is added
by modernish on the others, fulfills their function well (with the exception
of 'ndiv', but it is rarely needed).

New modules:
	var/arith/cmp	arithmetic comparison: eq,ne,gt,ge,lt,le
	var/arith/ops	arithmetic operators: inc,dec,mult,div,ndiv
	var/arith	both of these

The rest of modernish is adapted to either use 'let' instead of the above or
use these modules.