die(): stop command execution on interactive yash and mksh
As a fallback for yash and mksh, send SIGINT to the main shell, which is
equivalent to pressing Ctrl+C. This works on every shell, but has certain
disadvantages: sending SIGINT causes an exit status that reflects that
Ctrl+C was pressed, and the signal might be trapped and ignored. Unsetting
the trap before sending the signal is possible, which would eliminate the
problem of the signal being ignored, but it would not be possible to restore
the trap after returning to the command prompt, which would make die()
destructive. For this reason I'm keeping the hacks that work on shells other
than yash and mksh.