mktemp(): faster parallel processing
Using $RANDOM turns out insufficient for a process that's repeatedly invoked
in a subshell because the RANDOM seed is not updated in the main shell, so
each subshell will get the same "random" number. Of course shells without
RANDOM already had this problem anyway since they only had $$ (PID of main
shell) to go on. So multiply that with the PID of the current subshell
(there is a trick to obtain this) to get a different suffix each time. This
reduces the load on parallel creation of temporary files, e.g.
mktemp-stresstest.sh is drastically sped up.

Also, convert the suffix to hexadecimal to somewhat reduce the length.