New fatal bug test FTL_SUBSHEXIT (incorrect exit status in subshells)
bash 4.3 and 4.4 compiled with '--enable-minimal-config' are broken when
executing commands in subshells: commands that exit unsuccessfully (exit
status > 0), such as 'false', yield exit status 0 within a subshell.
For instance:
	( false && echo true || echo false )
outputs 'true' with this bug.

On bash 4.3 and 4.4 compiled with both '--enable-minimal-config' and
'--enable-alias' the subshell works correctly once, and then the bug shows
up. The location of this test in bin/modernish ensures its detection even
under this cirumstance, as it's done after other subshells have already been
executed.

Also, tweak FTL_UNALIASA and FTL_NOALIAS.