VERSION 0.3.0: 'use' modules from hashbang comments (ksh93 compat)
Until now, AT&T ksh93 was not usable as the default shell for use with the
generic "#! /usr/bin/env modernish" hashbang path. The problem was that
aliases defined in modules didn't expand, so modules that depend on them
(such as everything in loop/) didn't work. That's why install.sh would
refuse to let you choose it as the default shell.

However, aliases from the main bin/modernish did always work on ksh93. That
made me think aliases from modules might work on ksh93 if bin/modernish
pre-sources the modules *before* sourcing the main program. This does indeed
turn out to be the case.

Consequently, this commit introduces a new way of using modules, collateral
to the normal one. The new way, required for use with the generic hashbang
path "#! /usr/bin/env modernish" if the script is to be compatible with
ksh93, and optional otherwise, goes like this:

	#! /usr/bin/env modernish
	#! use safe -w BUG_UPP -w BUG_APPENDC
	#! use someothermodule
        normal commands start here

All lines that (a) start with "#!" and (b) IMMEDIATELY follow the main
hashbang path, are evaluated by bin/modernish. If anything else goes in
between, even an empty line, the rest of the hashbang lines will be ignored.
The part after the "#!" is taken as a shell command and executed as such.

This means such lines may in fact contain arbitrary shell commands, which
would be executed before the main program is sourced. It's not clear what
actual use other than pre-sourcing modules this could possibly have, but
allowing arbitary commands there also carries no harm that I can think of.

NOTE: the new hashbang method is not supported in modules. If a module
depends on another module, the normal 'use' command must be invoked.