pop(): better BUG_ARITHTYPE workaround
On zsh with BUG_ARITHTYPE, variables acquire an arithmetic-only content
restriction after arithmetic assignment, even under 'emulate sh', which is
contrary to POSIX which prescribes a typeless shell language. Such a
variable needs to be unset before it can be re-created to have arbitrary
values again. This requires a workaround version for pop() as the value to
be popped is an arbitrary string.

Previously the workaround version would always unset the variable before
popping it. However, this has the unfortunate side effect of unexporting a
variable that has its export bit set, which can lead to undesirable side
effects (what if that variable is PATH? Turns out zsh doesn't allow PATH to
be unexported, but the point stands for other exported variables.)

So now, with this commit, it is first tested whether the variable has an
arithmetic type restriction before unsetting it. This is done by simply
assigning the name of the stack pointer variable and checking if the
resulting value is the variable's name (as expected) or its value. While
this does not guarantee that an exported variable will never be unexported,
hopefully this minimises that side effect.