New $SIGPIPESTATUS sys constant instead of _Msh_PIPEstatus
Now always init the variable used to store the exit status corresponding to
a process being killed by SIGPIPE, and make it an official modernish system
constant $SIGPIPESTATUS. This is done because harden() is no longer the only
function that needs this value; traverse() will use it (as of a subsequent
commit) and it is projected that other functions might need it in future.

This commit also makes the test procedure more robust, checking against
shell bugs. Bash 2 is explicitly blacklisted for the faster method because
its 'kill' builtin has a race condition that can cause it to kill the main
shell instead of the subshell. (A proper bug test for this is not feasible;
the problem is intermittent.)

WRITEUP.md: Document $SIGPIPESTATUS; move documentation for all modernish
system constants from comments in bin/modernish into here; other tweaks.

SHELLBUGS: document a relevant bug in various ksh93 versions; other tweaks.

(Note: the name PIPESTATUS was tried first, but it conflicts with a
read-only PIPESTATUS array on mksh.)