mktemp(): autocleanup requires -s
Setting a trap for autocleanup does not work from a subshell, so mktemp's -C
([C]leanup) option can't be used if mktemp is called in the usual way, by
forking a command substitution subshell. This can't work:
	tmpfile=$(mktemp -C /tmp/somefile.XXXXXX)
because the cleanup trap would be executed immediately upon exiting the
command substitution subshell, rendering the whole thing an expensive
non-operation. Subshells can't influence the main shell anyway.

To avoid unexpected problems, ban the use of mktemp in subshells (including
command substitution) in the best way possible: require the -s option for no
writing to standard output, so command substitution is useless. The output
is stored in $REPLY instead, and reading it requires executing mktemp() in
the main shell.

(In future we'll have a function insubshell() that determines if we're in a
subshell and mktemp() will use that instead for more accurate blacklisting
of the -C option.)

Also, tweaks to the option parser (which is now automatically generated
using a soon-to-be-published parser generator module) and the comments.