Downgrade BUG_BRACSQBR to fatal FTL_BRACSQBR
FreeBSD sh was the only current shell with BUG_BRACSQBR. This bug has now been
fixed in the release version, so we can downgrade it to a fatal bug and stop
worrying about it.

A small range of AT&T ksh93 versions (version JM 93t+ 2010-??-??) also has
this bug, at least versions "JM 93t+ 2010-03-05" and "JM 93t+ 2010-06-21".
Older and newer versions don't have it. It just doesn't seem worth it to
keep working around the bug just for these.

As a result, $SHELLSAFECHARS can now contain ']' (but not '[' as this may
introduce a glob pattern).

Marking this bug fatal also enables the use of $ASCIICHARS with 'case' for
checking if a variable contains a non-ASCII character. (This is a readonly
variable containing the complete ASCII set that is in the pipeline for the
next commit).

bin/modernish:
- add FTL_BRACSQBR as a fatal bug test
- group the three fatal 'case' bugs together
- add ']' to $SHELLSAFECHARS
- sort the non-alphanumeric characters in $SHELLSAFECHARS to something
  closer to ASCII order, with BUG_BRACQUOT problem characters ^, ! and -
  conspicuously at the end

libexec/modernish/cap/BUG_BRACSQBR.t:
- removed

README.md:
- BUG_BRACSQBR entry removed