New string ops: toupper(), tolower()
The new string operation functions 'toupper' (convert a variable to upper
case letters) and 'tolower' (same for lower case) are added to the main
bin/modernish library and not a module, because they are needed in the main
library.

Modernish does extensive sanity checks and tries very hard to make these
functions as efficient as possible. If `typeset -u` and `typeset -l` are built
in to shell, they are used -- either for converting ASCII only or for
converting UTF-8, depending on their capability. Otherwise the functions fall
back to an external command, which modernish chooses out of 4 possibilities, in
hopes at least one of them supports UTF-8 properly. If all options fail and no
external utilities support UTF-8, BUG_CNONASCII is detected. This bug is only
tested for in UTF-8 locales,

Other tweaks:
- Add $CC7F (DEL) control character constant.
- Add $ASCIIUPPER and $ASCIILOWER constants (the ASCII alphabet).
- Add $ASCIICHARS constant (the complete ASCII character set, 1 to 127).