builtin(): call die() if command is not a builtin
Since scripts have "thisshellhas --bi=COMMAND" at their disposal to check for
a builtin, there's no good reason to treat an attempt to call 'builtin' with
a non-builtin as anything but a fatal error.

There is a precedent of sorts: mksh exits the shell if 'builtin' is called
with a non-builtin.