yash 'return' bug workaround for interactive use
Yash 2.43 does not recognise 'return' in dot scripts (outside of
functions) if the dot script is sourced from the interactive command
line, so doing '. modernish' from the command line in yash caused the
'return' to be ignored with an error message. Subsequent code that
should not have been executed then exited the shell. As a workaround,
send SIGINT to the main shell if 'return' fails.

Ref.: https://osdn.net/ticket/browse.php?group_id=3863&tid=36766