Fix trap stack for sig names w non-varname chars
The trap stack was not working for signal names containing '+' and
'-', which occur on Linux (the "real-time signals" like RTMIN+2,
RTMAX-3 etc.). Fix this by keeping two variables instead of one:
one for the signal name and one for the corresponding variable
name elements. This required some significant reworking of trap
stack-related functions. Signal names that contain characters that
are invalid for variable names are converted to valid variable name
elements with 'tr', where lowercase letters (which can never be
part of a signal name) are used instead of the special characters.