temp files: no more 'set -C', use mkdir/ln instead
The technique to create a uniquely named file with 'set -C' ('set
-o noclobber') and trying output redirection with different names
until it succeeds is not as safe as I thought it was:

- Shell implementations of 'set -C' are not guaranteed to be
  atomic, opening the possibility of race conditions.

- Nothing prevents output redirection under 'set -C' from writing
  into a FIFO or other device file; this can also be used by an
  attacker.

Instead, to create a file, first create a directory with 'mkdir'
(which is safe and atomic), then create the file inside that
subdirectory, and then hardlink it back into the parent directory
with 'ln' (which is also safe and atomic), then remove the
directory.

This method of course has worse performance but should avoid
vulnerabilities. In general, creating temporary directories is best
anyway.

These changes apply to:
- mktemp() in libexec/modernish/sys/base/mktemp.mm
- the bug test in libexec/modernish/cap/BUG_APPENDC.t
- the trap stack implementation of the 'trap' command to print the
  traps, in bin/modernish

Ref.:
https://www.mail-archive.com/austin-group-l@opengroup.org/msg00435.html
https://www.mail-archive.com/austin-group-l@opengroup.org/msg00440.html