insubshell(): use BUG_FNSUBSH to make it actually work on ksh93!
A correctly functioning insubshell() function (to test if we're
currently in a subshell) is essential for several other modernish
functions (i.e. the trap stack, die(), harden(), and mktemp()).

On AT&T ksh93 as of 2010-ish, the combination of BUG_KSHSUBVAR (the
${.sh.subshell} counter is broken) and the non-forking subshell
feature (i.e. non-background subshells share the main shell's PID)
makes it impossible to test in any normal way if we're in a
subshell; we can neither use that counter nor use the canonical
$(sh -c 'echo $PPID') == $$ trick.

But, in a demonstration of the idea that there really is no
technical difference between a bug and a feature, ksh93 does have
one bug/feature we can use: BUG_FNSUBSH makes it impossible to
unset a shell function in a non-background subshell. The 'unset -f'
command silently fails and the function remains set. We can easily
(ab)use this to test if we're in a non-background subshell.

bin/modernish:
- Overhaul insubshell() by storing platform-specific code in
  variables and piecing together the function from those.
- Now have two AT&T ksh93 versions of insubshell(): one for
  ksh93 versions without BUG_KSHSUBVAR (those are the 1993-ish
  ones) and one for those with. The non-bug version simply checks
  if ${.sh.subshell} is zero. The bug version unsets a dummy
  function and checks if it's actually unset.
- All versions of ksh93 bizarrely reset the subshell counter to
  zero in background job subshells, and BUG_FNSUBSH doesn't apply
  in those either. But since those *do* fork new processes, and
  "$$" remains the PID of the main shell, simply use the canonical
  "compare the PIDs" trick.
- Hard-fail modernish init if we ever detect a ksh93 version with
  BUG_KSHSUBVAR but without BUG_FUNSUBSH; we have no way left to
  compensate and things *will* break. (There currently isn't any
  such version.)
- Remove the old, brittle, insufficient BUG_KSHSUBVAR workaround
  from harden() as it's now superfluous.

libexec/modernish/tests/insubshell.t (new file):
- Test suite script for insubshell().

libexec/modernish/tests/run.sh:
- Use sys/base/mktemp so tests can safely use temporary files.
- To make tracing failed subshell tests in insubshell.t possible,
  no longer run the trace of a failed test in a subshell.

install.sh:
- To make the subshell tests in insubshell.t possible, no longer
  run the test suite in a subshell from the installer shell, but
  launch a new shell to run it instead.

libexec/modernish/cap/BUG_KSHSUBVAR.t:
- Update comments.