FTL_FLOWCORR: zsh on Solaris has fatal program flow corruption bug
This is a zsh bug that *only* manifests on Solaris, at least
version 11.3.

If a subshell exits due to an error in a special builtin or redirection,
execution flow is corrupted in such a manner that, when end of file is
reached without an explicit 'return' or 'exit' being encountered,
execution of the file does not end but restarts at the point exactly
after the subshell was exited. The second time around, if program
specifics allow it, execution ends normally.

The bug only manifests if POSIXBUILTINS is active, and only on
Solaris, but on *all* zsh versions to date.

Though it's a fatal bug that needs to be tested for at init time
(so modernish can refuse to launch if found), due to the nature of
the bug it needs it's own dot script, so it can reach end of file
there without 'return' or 'exit' being invoked, which is part of
what triggers the bug.

bin/modernish: zsh-specific settings init:
- Source the test script, but only if we're on zsh. This bug is so
  unlikely to turn up anywhere else it'd be a pointless waste of
  cycles to run it on other shells.

libexec/modernish/cap/aux:
- New directory for helper (= auxiliary) scripts that need to be
  sourced from either bin/modernish or one of the feature test
  scripts.

libexec/modernish/cap/aux/FTL_FLOWCORR.t:
- The bug test dot script. See the comments there for how it works.