die() now kills all subprocesses, reliable emergency program halt
One of modernish's essential design principles is that shell
programs should never continue to run when it has been determined
to be in an inconsistent state. So a reliable die() that kills the
program, including all its subshells, even when invoked *from* a
subshell, is a must-have.

So far, die() was using 'pgrep -P' or a fallback with 'ps' and
'awk' to find the shell's direct subprocesses. This commit abandons
'pgrep' and uses an 'awk' program that recurively finds all the
shell's subprocesses from the 'ps' output, resulting in a reliable
emergency program halt on all shells (with no orphan background
processes wreaking havoc on the terminal or your storage).