QRK_LOCALUNS2: 'unset' doesn't unset in subscope of local scope (bash)
New quirk QRK_LOCALUNS2 found, which is a more treacherous version
of QRK_LOCALUNS that is unique to bash. The 'unset' command works
as expected when used on a local variable in the same scope that
variable was declared in, HOWEVER, it makes local variables global
again if they are unset in a subscope of that local scope, such as
a function called by the function where it is local.

As with QRK_LOCALUNS, this means 'unset' will not necessarily unset a
variable if local variables are used!

Ref.: the bug-bash thread starting at
http://lists.gnu.org/archive/html/bug-bash/2017-03/msg00105.html