BUG_KUNSETIFS: another obscure corner-case bug on AT&T ksh93
Can't unset IFS on ksh93 under very specific circumstances.

'unset IFS' or 'unset -v IFS' is a known POSIX shell idiom to
activate default field splitting. With this bug, the 'unset'
builtin silently fails to unset IFS if all four of the following
conditions are met:

  1. IFS is set and empty in the main shell (split is disabled).
  2. We're in a subshell.
  3. We're executing an 'eval' or a trap.
  4. If we're executing an 'eval', the code 'eval' is executing
     results from a parameter expansion.

Re 2, it makes no difference whether the subshell is within the
'eval' or the 'eval' is within the subshell.

The bug is not triggered for any other variable -- just IFS.

Workaround: assign anything to IFS (even the empty value that was
already there) immediately before unsetting it. This makes 'unset'
work again.

The trap stack is going to need this workaround when we restore the
state of IFS before executing a trap. The conditions above are met
when running _Msh_doOneStackTrap(). See next commit.

(See also BUG_IFSISSET on ksh93: you can't test in any normal
way if IFS is set or not. But modernish isset() is safe, as it
has a workaround for that.)