trap stack: add 'pushtrap --nosubshell' (advanced usage)
Mostly, running stacked traps in one subshell each is a really good
idea: it ensures one trap can't interfere with the next (e.g. by
changing IFS/glob, exiting, etc). However, sometimes it may be
necessary to restore the main shell's global settings on receiving
a signal or encountering an error -- particularly if you're on an
interactive shell where a signal or die() will not kill the shell.

The --nosubshell option can now be used to push a signal trap for
running in the main shell. This should be considered advanced
usage, hence the long-form option name; we do not want this to
become an automatism.

bin/modernish:
- Add --nosubshell option to pushtrap().
- Rework the relevant trap stack functions to deal with the new
  functionality.

libexec/modernish/tests/trap.t:
- Verify --nosubshell by checking for survival of a variable.

libexec/modernish/var/setlocal.mm:
- On interactive shells, while executing a setlocal block, push a
  nosubshell SIGINT trap to restore global settings. This fixes the
  bug that the shell settings could be left in an inconsistent
  state when pressing ^C while executing a setlocal block, or if
  die() was invoked (which issues SIGINT on interactive shells).

README.md:
- Document the new --nosubshell option for pushtrap().