fix: detect BUG_CMDSPEXIT properly
BUG_CMDSPEXIT should be the bug ID for any failure to stop a POSIX
"special builtin" from exiting the shell on error when prefixed
with 'command'. However, some variants on bash <= 4.0, and all
current mksh versions weren't caught. This should fix that.

'command eval' is exempt from this because using it to test for
syntax errors in the main shell causes too many shells to either
exit or become crash-prone. So is 'command exec' which still causes
every shell to exit if the file is not found. Neither should ever
be tested outside of a test subshell.

libexec/modernish/cap/BUG_CMDSPEXIT.t:
- Test a more fully representative sample of special builtins to
  hopefully catch all the shells with this bug.

libexec/modernish/tests/util.t:
- Update regression test 3 to test all POSIX special builtins
  except 'exec', 'eval', 'exit' and 'return', so any failure to
  detect BUG_CMDSPEXIT properly is caught here.

README.md:
- Update & correct the BUG_CMDSPEXIT doc.