BUG_TRAPSUB0: subshells executed by traps always have exit status 0
Some shells have this bug where if you test anything ( in a
subshell ) within a trap, it will always appear to succeed because
subshells exit with status 0 even if the last command they executed
had status >0 or exited the shell with an error.

This broke modernish 'isset -r' (test if variable is read-only)
when used in a trap action, because that test is done by trying to
assign a value to the variable in a subshell.

The bug is triggered differently on different shells:

- bash<=4.0: only for exiting on errors (e.g. assigning to readonly)
- dash 0.5.9-0.5.10.2: consistent for signals; intermittent for EXIT
  Ref.: https://www.spinics.net/lists/dash/msg01750.html
- yash <= 2.47: consistent, EXIT traps only
  Ref.: https://osdn.net/projects/yash/ticket/38774

But the workaround is always the same: use an explicit 'exit' from
the subshell, so 'exit $?' to pass down the exit status from the
last command. Thus, these bugs can go under one ID.

This won't work if a special builtin fails, though, because the
shell will exit immediately and will never get to your 'exit'
command. Thankfully POSIX provides the 'command' builtin to disable
this behaviour. But that doesn't work on shells with BUG_CMDSPEXIT.

Thankfully, these two bugs are not known to coexist on any
particular shell out there in the wild. Just to make sure, check
for BUG_TRAPSUB0 + BUG_CMDSPEXIT as a fatal bug combination in the
sanity checks.

libexec/modernish/cap/BUG_TRAPSUB0.t:
- Added.

libexec/modernish/tests/trap.t:
- Add regression test for correct detection of this bug.

bin/modernish:
- Change 'isset -r' to work in traps on shells with BUG_TRAPSUB0.

libexec/modernish/tests/isset.t:
- Add regression test for 'isset -r' within a trap.

libexec/modernish/tests/@sanitychecks.t:
- Fail on detecting BUG_TRAPSUB0 & BUG_CMDSPEXIT on the same shell.

README.md:
- Add BUG_TRAPSUB0 documentation.