unconditionally set useful PS4 for xtrace
If PS4 has a default value, now always set a useful PS4 for xtrace
output and not just if 'set -x' was on at initialisation time. This
way, if 'set -x' is activated after init, it will be useful.
Modernish has a *lot* of shell code so this is quite indispensable.

Depending on the shell, trace:
 * current time, or time elapsed -- with precision if available
 * either subshell PID (P) or subshell level (S)
 * current file (last 2 pathname elements)
 * current shell function
 * line number (L)
 * the previous command's exit status (e)

bin/modernish:
- Set PS4 at an early stage in init, right after initialising
  POSIX compliance for various shells.
- Add tracing of time and subshell PID/level.

libexec/modernish/tests/run.sh:
- Remove now-superfluous duplicate PS4 init.