v0.10.1dev: var/loop: make Solaris sh/ksh/bash work (slowly)
libexec/modernish/var/loop.mm:
- Solaris (at least up to 11.4) is strangely broken: FIFOs aren't
  ready for use immediately after creation. We have to wait at
  least half a second before our 'exec' will work, otherwise the
  shell will simply hang on the next 'read'. Use 'sleep 1' just to
  be safe.
- This applies to ksh93 (/bin/sh) and bash, but *not* to yash, zsh
  and mksh, which cope just fine. So we do ugly shell version
  detection so the latter three can work at normal speed.

This is obviously a very ugly hack and we need a real workaround!
Any Solaris experts out there?

Pre-creating the FIFOs and using them as needed would work around
the problem. Unfortunately, this is not an option. With a directory
of prepared FIFOs, it cannot be guaranteed that two modernish
programs (or even background processes of the same program) will
not access the same FIFO at the same time, corrupting loops in both
processes. We have to atomically create our own on the fly each
time to guarantee unique use.