die() et al: use 128, not 125, for non-signal fatal error status
The die() function used exit status 125 (one lower than 126, "can't
execute") for the current process if it doesn't SIGKILL it. However
upon re-reading POSIX, I'm noticing that, although 126 and 127 and
>128 are reserved for special purposes, 128 itself is not.

I also noticed Git uses 128 for fatal error. This works well with
the default trace() hardening recipe which kills the program if the
traced/hardened command has an exit status of 126 or higher.

Statuses starting from 129 correspond to termination by a signal,
that is, the signal number + 128. Signals start at 1. That makes
128 the logical choice for "non-signal fatal error" as (a) it
corresponds to pseudosignal 0, the EXIT trap in the shell, and
(b) in binary, it is the first number with the eighth bit set.

References:
http://pubs.opengroup.org/onlinepubs/9699919799/utilities/V3_chap02.html#tag_18_21_18
http://pubs.opengroup.org/onlinepubs/9699919799/utilities/V3_chap02.html#tag_18_08_02

bin/modernish:
- die(): non-interactive: unless using signal, exit 128 (not 125)
- die(): interactive: if SIGINT is ignored, return 128 (not 125)
- _Msh_doTraps(): interactive: SIGINT handling: idem

install.sh, uninstall.sh:
- exit 128, not 125, on failure to find a good POSIX shell

libexec/modernish/var/loop.mm:
- _loop_die(): exit loop generator with status 128, not 125