FTL_TILDSPLIT: cheaper way to block dash <= 0.5.6.1
Dash 0.5.6.1 and earlier have a program flow corruption bug
involving sourcing dot scripts with 'command .', which kills
modernish, so we need to block it. However, the related
FTL_FLOWCORR2 fatal bug test is expensive to run on every init as
it involves forking two processes. But there turns out to be
another dash <= 0.5.6.1 bug (fixed in 0.5.7) that can be called
fatal and is much cheaper to test for, so we can use that to block
these dash versions instead: tilde expansion doesn't work right.

bin/modernish:
- Remove expensive FTL_FLOWCORR2 fatal bug test.
- Add FTL_TILDSPLIT: tilde expansion is subject to field splitting.

libexec/modernish/cap/aux/FTL_FLOWCORR2.t:
- Removed.