part revert ee59622c: remove FTL_TILDSPLIT
Turns out that bug, which is hardly fatal in and of itself, blocks
too-recent Busybox ash versions that are otherwise perfectly
capable. So we can't really use it to block dash < 0.5.7. Since
there doesn't seem to be a non-expensive way to test for those dash
versions, we'll just have to be happy with those dash versions
having a couple of FAILs in the regression test suite.