shell options stack & trap stack: export LC_ALL=C to 'tr'
Solaris 'tr' does completely bizarre things in UTF-8 mode when
changing case using the classic idiom 'tr a-z A-Z'. Unfortunately
we can't change to 'tr [:lower:] [:upper:]' which works on Solaris
because Busybox doesn't support it. We are only converting
shell-safe characters anyway, so passing LC_ALL=C fixes it.