remove BUG_XTRCREDIR -- unclear if actually bug
It is unclear if it actually a bug if the xtrace is redirected to
standard output on something like 'set -x; command 2>&1' (which
causes problems if that happens within a command substitution; the
xtrace becomes part of the captured output). The behaviour of
various shells also differs in too may ways for BUG_XTRCREDIR to be
a useful bug/quirk detection test.

To avoid problems, it should be standard practice to turn off
xtrace before standard error is redirected to standard output
within a command substitution.