0.15.0-dev: new: sys/cmd/procsubst: portable process substitution
Process substitution (as in ksh, bash, zsh) is functionality
clearly missing from the POSIX shell. Reading from a here-document
containing a command substitution is not a complete equivalent as
the command is not executed in the background. So let's provide a
"poor man's" process substitution using FIFOs and an alternative
syntax that can be added to the POSIX shell. It should be
functionally equivalent.

lib/modernish/mod/sys/cmd/procsubst.mm,
share/doc/modernish/examples/procsubst.sh:
- Added.

README.md:
- Fix: add chapter heading for sys/cmd module directory.
- Document sys/cmd/procsubst.

bin/modernish:
- Version bump to 0.15.0-dev.