0.15.4dev: BUG_FORLOCAL; dot script for $DEFPATH detect/validate
BUG_FORLOCAL: a 'for' loop in a function makes the iteration
variable local to the function, so it won't survive the execution
of the function. Found on: yash. This is intentional and documented
behaviour on yash in non-POSIX mode, but in POSIX terms it's a bug,
so we mark it as such.

lib/modernish/cap/BUG_FORLOCAL.t:
- Added.

lib/modernish/tst/loop_cond.t:
- Add regression test for correct BUG_FORLOCAL detection.

lib/_install/defpath.sh:
- Added.

lib/_install/goodsh.sh:
- Work around BUG_FORLOCAL.

lib/_install/*.sh:
- 'return', don't 'exit' from dot scripts, so calling scripts can
  decide how to handle errors.
- Some old shells have problems 'return'ing directly from dot
  scripts, so work around that by wrapping them in temp shell
  functions.

bin/modernish:
- Version bump to 0.15.4dev.
- Avoid an interactive shell exiting if a non-installed modernish
  cannot determine $DEFPATH or a good shell. To that end, refactor
  and move _Msh_initExit() so we can use it early.
- Use lib/_install/defpath.sh to determine and validate DEFPATH.
- Move down ${_Msh_cap} init to a location that is more sensible
  now that we no longer store fatal bug IDs in it. (re: d45d4912)
- Return or exit with status 128 (kind of standard for 'fatal
  error') instead of 124 on initialisation feature.

README.md:
- Document BUG_FORLOCAL.