move builtin WRN_NOSIGPIPE cap ID to cap/WRN_NOSIGPIPE.t
While the initialisation of $SIGPIPESTATUS must be unconditional
and therefore remain in bin/modernish, that routine already gives
an indication if SIGPIPE is ignored: SIGPIPESTATUS is set to 99999.
So we can delegate the actual setting of the cap ID to a regular
cap/*.t script, for consistency.

lib/modernish/cap/WRN_NOSIGPIPE.t:
- Added.

bin/modernish:
- No longer add WRN_NOSIGPIPE to _Msh_cap manually.

README.md:
- No longer indicate WRN_NOSIGPIPE is a builtin detection test.
- Copy-edit its documentation.