move builtin RANDOM cap ID to cap/RANDOM.t
While the initialisation/seeding of $RANDOM must be unconditional
and therefore remain in bin/modernish, that routine already gives
an indication if the shell has no $RANDOM: it empties the variable
before unconditionally making it read-only.
So we can delegate the actual setting of the cap ID to a regular
cap/*.t script, for consistency.

lib/modernish/cap/RANDOM.t:
- Added.

bin/modernish:
- No longer add RANDOM to _Msh_cap manually.

README.md:
- No longer indicate RANDOM is a builtin detection test.