aux/fatal.sh: block ksh93 < 93u+ 2012-08-01; yash -o posix < 2.44
AT&T ksh93 before 2012 has a bug that causes a background subshell
to terminate instead of stop if it is issued SIGSTOP. That's the
kind of bug we can't easily detect in a test, but it also has a
subtle parser bug involving 'for' loops that we can use to block
pre-2012 ksh93 release versions. Those are now old & buggy enough
they're not worth supporting.

As a side effect we also block yash < 2.44 in POSIX mode, which has
a similar parser bug with 'for' loops.

Finally, this commit replaces the bug test that blocked dash <
0.5.7 with another one that doesn't come with the cost of a fork.

lib/modernish/aux/fatal.sh:
- Always test for shell grammar bugs using an 'eval'; much safer.
- Move the "Bugs with shell grammar" section to the top; there's
  now no reason for this to be near the bottom of the script.
- Add FTL_FORSCOLON, blocking ksh93 < 93u+ & yash -o posix < 2.44.
- Remove FTL_FLOWCORR2, which we used to block dash < 0.5.7, at the
  cost of forking a subshell in which to try to exec a nonexistent
  external command.
- Add FTL_TILDSPLIT, blocking dash < 0.5.7 without costing a fork.

README.md:
- Update 'Supported shells' section.